+++ src/layouts/loading/index.vue
@@ -5,18 +5,18 @@
 </template>
 
 <script lang="ts">
-import { defineComponent } from 'vue'
+  import { defineComponent } from 'vue'
 
-export default defineComponent({
-  name: 'LoadingComponent',
-})
+  export default defineComponent({
+    name: 'LoadingComponent',
+  })
 </script>
 
 <style lang="less" scoped>
-.loading-wrapper {
-  min-height: 50vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
+  .loading-wrapper {
+    min-height: 50vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
 </style>