+++ src/layouts/logo/index.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="logo-wrapper">
+  <div class="logo-wrapper" :style="{ 'background-color': bgColor }">
     <img v-if="showLogo" class="logo-img" src="../../assets/logo.png" />
     <div
       v-if="showTitle"
@@ -14,7 +14,7 @@
   import { computed, defineComponent } from 'vue'
   import { projectName } from '../../setting'
   import useAppConfigStore from '@/store/modules/app-config'
-  import { SideTheme } from '@/store/types'
+  import { SideTheme, ThemeMode } from '@/store/types'
   export default defineComponent({
     name: 'Logo',
     props: {
@@ -34,18 +34,22 @@
     setup() {
       const appStore = useAppConfigStore()
       const bgColor = computed(() => {
-        if (appStore.sideTheme === SideTheme.DARK) {
-          return 'var(--color-menu-dark-bg)'
-        }
-        if (appStore.sideTheme === SideTheme.WHITE) {
-          if (appStore.layoutMode !== 'ttb') {
+        if (appStore.layoutMode !== 'ttb') {
+          if (appStore.sideTheme === SideTheme.DARK) {
+            return 'var(--color-menu-dark-bg)'
+          }
+          if (appStore.sideTheme === SideTheme.WHITE) {
             return appStore.sideTheme === SideTheme.WHITE
               ? 'var(--color-white)'
               : appStore.sideTheme === SideTheme.DARK
               ? 'var(--color-menu-dark-bg)'
               : 'transparent'
           }
-          return 'var(--color-white)'
+          return 'transparent'
+        } else {
+          return appStore.theme === ThemeMode.DARK
+            ? 'var(--color-menu-dark-bg)'
+            : 'var(--color-white)'
         }
       })
       return {
@@ -62,8 +66,8 @@
     display: flex;
     justify-content: center;
     align-items: center;
+    color: var(--color-text-1);
     border-bottom: 1px dashed var(--color-border);
-    background-color: v-bind(bgColor);
     .logo-img {
       width: 30px;
     }