+++ src/views/index/components/Title.tsx
@@ -1,0 +1,14 @@
+import { defineComponent, toRefs } from 'vue'
+
+export default defineComponent({
+  props: {
+    title: {
+      type: String,
+      default: '',
+    },
+  },
+  setup(props) {
+    const { title } = toRefs(props)
+    return () => <div>{title}</div>
+  },
+})